SKELANA pscgrc/psclrc: fix C++ struct memory layout (was reading garbage)
Codex mppqrb2p caught a 3-bug stack in the C++ wrappers that meant
our RICH gas plot did NOT match DELPHI 1994 fig 7.19(b). After the
fix, gas THEG median (NPHG>3) = 0.0572 rad = exactly the expected
saturation 58 mrad ✓.

Bugs in pscgrc.hpp (gas RICH):
  1. kgric was declared [MTRACK][6] but Fortran has LENGRC=5
     → SHIFTED ALL subsequent arrays by MTRACK*sizeof(int) bytes
     → THEG/SIGG/NPHG/NEPG/FLAGG accessors all returned garbage
     → THIS is what made the gas plot impossible (values 5-30 mrad
       instead of ~58 mrad saturation)
  2. nepg was declared float; Fortran has INTEGER NEPG
  3. flagg was declared int; Fortran has REAL FLAGG (the flag bit
     pattern is encoded in float bits)

Fortran COMMON (per /cvmfs/.../stdcdes.car +KEEP,PSCGRC):
  INTEGER LENGRC; PARAMETER (LENGRC = 5)
  INTEGER NPHG, NEPG
  REAL    THEG, SIGG, FLAGG
  COMMON /PSCGRC/ NGRIC, KGRIC(LENGRC,MTRACK),
 +                THEG(MTRACK), SIGG(MTRACK),
 +                NPHG(MTRACK), NEPG(MTRACK),
 +                FLAGG(MTRACK)

Bugs 2,3 also in psclrc.hpp (liquid RICH) — fixed there too. Liquid
plot mostly worked before only because its LENLRC was already 5
(no array shift), and π saturation dominated; the wrong NEPL type
was harmless for plotting, FLAGL type didn't matter for the angle.

+++ delphi-analysis/include/skelana/psclrc.hpp
@@ -19,28 +19,41 @@ namespace skelana
 *     QLRIC(4,I) - NEPL , Expected number of photoel. in liquid
 *     KLRIC(5,I) - FLAGL, RICH sector flag in liquid
 *
+* Fortran COMMON layout (stdcdes.car +KEEP,PSCLRC):
+*   INTEGER NPHL, NEPL                         <- NEPL is INTEGER
+*   REAL    THEL, SIGL, FLAGL                  <- FLAGL is REAL
+*   COMMON /PSCLRC/ NLRIC, KLRIC(LENLRC,MTRACK),
+*  +                THEL(MTRACK), SIGL(MTRACK),
+*  +                NPHL(MTRACK), NEPL(MTRACK),
+*  +                FLAGL(MTRACK)
 *
+* PRIOR BUG (Codex mppqrb2p): nepl declared float but Fortran has it
+* as INTEGER; flagl declared int but Fortran has it as REAL.
+* Liquid plot still mostly worked because π saturation dominates and
+* the per-track THEL/SIGL/NPHL slots happened to land in the right
+* memory region even with the wrong types; gas was much more sensitive
+* to the layout drift via PSCGRC's extra [6] column bug.
 */
     inline const int LENLRC = 5;
     extern "C" struct
     {
-        int nlric;
-        int klric[MTRACK][LENLRC];
+        int   nlric;
+        int   klric[MTRACK][LENLRC];
         float thel[MTRACK];
         float sigl[MTRACK];
-        int nphl[MTRACK];
-        float nepl[MTRACK];
-        int flagl[MTRACK];
+        int   nphl[MTRACK];
+        int   nepl[MTRACK];      // INTEGER (was float — bug)
+        float flagl[MTRACK];     // REAL (was int — bug)
     } psclrc_;
 
     inline int &NLRIC = psclrc_.nlric;
-    inline int &KLRIC(int i, int j) { return psclrc_.klric[j - 1][i - 1]; }
+    inline int   &KLRIC(int i, int j) { return psclrc_.klric[j - 1][i - 1]; }
     inline float &QLRIC(int i, int j) { return *reinterpret_cast<float*>(&psclrc_.klric[j - 1][i - 1]); }
-    inline float &THEL(int i) { return psclrc_.thel[i - 1]; }
-    inline float &SIGL(int i) { return psclrc_.sigl[i - 1]; }
-    inline int &NPHL(int i) { return psclrc_.nphl[i - 1]; }
-    inline float &NEPL(int i) { return psclrc_.nepl[i - 1]; }
-    inline int &FLAGL(int i) { return psclrc_.flagl[i - 1]; }
+    inline float &THEL (int i) { return psclrc_.thel[i - 1]; }
+    inline float &SIGL (int i) { return psclrc_.sigl[i - 1]; }
+    inline int   &NPHL (int i) { return psclrc_.nphl[i - 1]; }
+    inline int   &NEPL (int i) { return psclrc_.nepl[i - 1]; }
+    inline float &FLAGL(int i) { return psclrc_.flagl[i - 1]; }
 } // namespace skelana
 
 #endif // SKELANA_PSCLRC_HPP
